moved webpack out of forge config as it was getting pretty crowded in that file
